#17: EventManager split onto base trait and implementation
