+++ src/test/scala/org/fs/mael/core/list/DownloadListSerializerSpec.scala
@@ -11,11 +11,10 @@ import org.fs.mael.core.backend.BackendManager
 import org.fs.mael.core.entry.BackendSpecificEntryData
 import org.fs.mael.core.entry.DownloadEntry
 import org.fs.mael.core.entry.LogEntry
-import org.fs.mael.core.event.EventManager
 import org.fs.mael.test.StubBackend
+import org.fs.mael.test.StubEventManager
 import org.fs.mael.test.TestUtils._
 import org.junit.runner.RunWith
-import org.scalatest.BeforeAndAfter
 import org.scalatest.FunSuite
 import org.scalatest.exceptions.TestFailedException
 
@@ -25,7 +24,7 @@ import com.github.nscala_time.time.Imports._
 class DownloadListSerializerSpec
   extends FunSuite {
 
-  val eventMgr = new EventManager
+  val eventMgr = new StubEventManager
 
   val backendMgr = (new BackendManager).withCode { backendMgr =>
     backendMgr += (new StubBackend, Int.MinValue)
